refactored fisher information matrix
