+++ batchglm/train/tf/base_glm_all/__init__.py
@@ -1,4 +1,5 @@
 from .estimator import EstimatorAll
 from .estimator_graph import EstimatorGraphAll
+from .fim import FIMGLMALL
 from .jacobians import JacobiansGLMALL
 from .hessians import HessianGLMALL
